Major bigfix needed.
Currently if I train a model and suppose that leads to a model with peak score X.
Now if I reload that model and run an initial evaluation in the fine tuning script then I get different scores at step 0.

What does this mean? Am I getting spurious dropouts?
I suspect that this is in multigpu mode.
What gives?

TODO: FIX THIS!